* develop: (68 commits)
  fix safe collect in profile
  add test for Get Movies By Genre Ids Use Case
  fix issues in based on true events and cinematic masterpieces
  fix base view model safe collect
  refactor collection list section ui on click
  Update domain/media/src/test/kotlin/com/giraffe/media/collections/usecase/ClearCollectionsCacheUseCaseTest.kt
  add six strategy to featured collection
  fix test coverage verification
  handle clear collections after logout
  create use case to get movies and series by sort
  create use case to get movies by keywords id
  create use case to get series by keywords id
  delete get filtered use case
  delete discover series function in repo, and add getSeriesByKeywordsId, getSeriesBySort
  delete get filtered use case
  delete discover movies function in repo, and add getMoviesByKeywordsId, getMoviesBySort
  handle clear collections after logout
  refactor late night strategy
  handle getting collections as flow
  - create use case to get series by list of genres - create use case to get series by list of genres - remove use case filter series - remove use case filter movies
  ...

# Conflicts:
#	presentation/home/src/main/java/com/giraffe/presentation/home/screen/show_more/ShowMoreStrategy.kt
#	repository/media/src/main/java/com/giraffe/media/movie/MovieRepositoryImpl.kt

+++ repository/media/src/main/java/com/giraffe/media/collections/datasource/remote/dto/CollectionItemDto.kt
@@ -10,7 +10,7 @@ data class CollectionItemDto(
     @SerialName("overview")
     val description: String? = null,
     @SerialName("vote_average")
-    val rating: Double? = null,
+    val rating: Float? = null,
     @SerialName("poster_path")
     val posterPath: String? = null,
     @SerialName("backdrop_path")
